+++ src/app/@modal/(.)book/[isbn]/loading.tsx
@@ -1,0 +1,7 @@
+import BookDetailLoading from '@/app/book/[isbn]/loading';
+
+const Loading = () => {
+  return <BookDetailLoading />;
+};
+
+export default Loading;
