+++ src/app/book/[isbn]/loading.tsx
@@ -1,7 +1,11 @@
 import { BookDetails } from './_components';
-
+import styles from './page.module.scss';
 const BookDetailLoading = () => {
-  return <BookDetails.Skeleton />;
+  return (
+    <div className={styles.container}>
+      <BookDetails.Skeleton />
+    </div>
+  );
 };
 
 export default BookDetailLoading;
