+++ src/app/book/[isbn]/page.tsx
@@ -4,7 +4,7 @@ import { BookDetails, RelatedBooks } from './_components';
 import styles from './page.module.scss';
 
 const BookDetailsPage = async () => {
-  await new Promise((resolve) => setTimeout(resolve, 100));
+  await new Promise((resolve) => setTimeout(resolve, 1000));
   return (
     <div className={styles.container}>
       <BookDetails.Loaded bookDetailData={BOOK_DETAIL_MOCK_DATA} />
