+++ src/components/overlay/index.tsx
@@ -1,1 +1,3 @@
 export { default as Toast } from './Toast';
+export { default as FullScreenModal } from './FullScreenModal';
+export { default as Portal } from './Portal';
